+++ lib/CHANGELOG.md
@@ -1,3 +1,7 @@
+# [unreleased]
+
+ * Purged old deprecated stuff
+
 # 1.4.2 (December 15th, 2020)
 
  * Moved CI to github actions